Shuffle, rename, externalize, and add tests for eth_getProof return types
