+++ src/chains/ethereum/utils/src/types/account-proof.ts
@@ -1,0 +1,17 @@
+import { Data, Quantity } from "@ganache/utils";
+
+export type StorageProof = {
+  key: Data;
+  proof: Data[];
+  value: Quantity;
+};
+
+export type AccountProof = {
+  address: Data;
+  balance: Quantity;
+  codeHash: Data;
+  nonce: Quantity;
+  storageHash: Data;
+  accountProof: Data[];
+  storageProof: StorageProof[];
+};
